+++ apps/www/pages/launch-week/14/index.tsx
@@ -2,8 +2,7 @@ import { NextSeo } from 'next-seo'
 import { useRouter } from 'next/router'
 import { useState } from 'react'
 import { Session } from '@supabase/supabase-js'
-import { LwCanvas } from '~/components/LaunchWeek/14/TicketCanvas'
-import TicketCanvas from '~/components/LaunchWeek/14/TicketCanvas2'
+import TicketCanvas from '~/components/LaunchWeek/14/TicketCanvas'
 import DefaultLayout from '~/components/Layouts/Default'
 import SectionContainer from '~/components/Layouts/SectionContainer'
 import { TicketState, ConfDataContext, UserData } from '~/components/LaunchWeek/hooks/use-conf-data'
